+++ validator/accounts/v2/BUILD.bazel
@@ -35,8 +35,8 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//shared/testutil:go_default_library",
-        "//validator/accounts/v2/iface:go_default_library",
         "//validator/keymanager/v2:go_default_library",
+        "//validator/keymanager/v2/direct:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
 )
