--- validator/accounts/v2/iface/BUILD.bazel
@@ -1,11 +1,0 @@
-load("@prysm//tools/go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["interface.go"],
-    importpath = "github.com/prysmaticlabs/prysm/validator/accounts/v2/iface",
-    visibility = [
-        "//validator:__pkg__",
-        "//validator:__subpackages__",
-    ],
-)
