+++ validator/keymanager/v2/direct/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
     ],
     deps = [
         "//shared/bls:go_default_library",
-        "//validator/accounts/v2/iface:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
 )
